requirements.txt
xlrd==1.1.0
itsdangerous==1.1.0
Flask==0.12.4
Flask_Bcrypt==0.7.1
Flask_WTF==0.14.2
WTForms==2.2.1
Flask_Login==0.4.1
Flask_SQLAlchemy==2.3.2
Pillow==5.3.0
python_dateutil==2.7.5
gunicorn==19.7.1
